SPU2: aligned bounce for savestate DataBlock; unsigned ADSR shift
FreezeIt/ThawIt were handed DataBlock references cast straight into
the frontend's savestate blob, which carries no alignment guarantee
while DataBlock embeds alignas(64) V_Core - 34 UBSan reports across
freeze/thaw.  Wrap both entry points: run in place when the blob
happens to be aligned, otherwise bounce through an aligned heap temp,
falling back to in-place access if the temp allocation fails (x86
tolerates it; beats losing the savestate).  Also shift ADSR p.Step in
the unsigned domain (left shift of negative value, same bits).

Savestate serialize/unserialize round-trip verified under ASan+UBSan
with byte-identical post-load framebuffer hashes.

+++ pcsx2/SPU2/defs.h
@@ -221,7 +221,8 @@ static __fi bool ADSR_Calculate(V_ADSR &v)
 	s32 counter_shift = p.Shift - 11;
 	s32 level_shift   = 11 - p.Shift;
 	u32 counter_inc = 0x8000 >> (counter_shift > 0 ? counter_shift : 0);
-	s32 level_inc   = p.Step  << (level_shift   > 0 ? level_shift   : 0);
+	/* p.Step can be negative; shift in the unsigned domain (same bits). */
+	s32 level_inc   = (s32)((u32)p.Step << (level_shift > 0 ? level_shift : 0));
 
 	if (p.Exp)
 	{
